% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VioplotQCForCluster.R
\name{VioplotQCForCluster}
\alias{VioplotQCForCluster}
\title{Fviolin plot of QC metrics}
\usage{
VioplotQCForCluster(rca.obj, prefix_name = "1")
}
\arguments{
\item{rca.obj}{RCA object.}

\item{prefix_name}{prefix filename.}
}
\description{
Fviolin plot of QC metrics
}
